Only restart hubot on master
PR's should not influence hubot, so this makes sure the script is only
called on non-PR's and on master.

Another annoying side effect is that the `after_success` hook will be
called 4 times (once for each element of the build matrix), at the
moment they don't have a solution for this [0], but since it's not that
big of a problem to have hubot restarted 4 times, I'll just act as if
this is not a real problem. #laikaboss

[0] https://github.com/travis-ci/travis-ci/issues/929